De-duplicate roundType localizing code
